+++ src/main/scala/osgifelix/BundleLocation.scala
@@ -3,7 +3,7 @@ package osgifelix
 import java.io.File
 import java.net.URI
 
-import org.apache.felix.bundlerepository.{Capability, Resource}
+import org.apache.felix.bundlerepository.{Resolver, Reason, Capability, Resource}
 
 /**
  * Created by jolz on 26/08/15.
@@ -20,9 +20,9 @@ object BundleLocation {
 case class BundleLocation(file: File, location: String)
 
 object ResolvedBundleLocation {
-  def apply(file: File):ResolvedBundleLocation = ResolvedBundleLocation(BundleLocation(file), true)
-  def apply(resource: Resource):ResolvedBundleLocation = ResolvedBundleLocation(BundleLocation(new File(URI.create(resource.getURI))),
-    resource.getCapabilities.exists(_.getName == Capability.FRAGMENT))
+  def apply(file: File):ResolvedBundleLocation = ResolvedBundleLocation(BundleLocation(file), true, None)
+  def apply(resolver: Resolver, resource: Resource):ResolvedBundleLocation = ResolvedBundleLocation(BundleLocation(new File(URI.create(resource.getURI))),
+    resource.getCapabilities.exists(_.getName == Capability.FRAGMENT), Some(resource, resolver.getReason(resource).toSeq))
 }
 
-case class ResolvedBundleLocation(bl: BundleLocation, fragment: Boolean)
+case class ResolvedBundleLocation(bl: BundleLocation, fragment: Boolean, resource: Option[(Resource, Seq[Reason])])